+++ public/components/common/hocs/error-boundary/with-error-boundary.tsx
@@ -1,5 +1,5 @@
 /*
- * Wazuh app - Module to export all the controllers
+ * Wazuh app - React HOCs handles rendering errors
  * Copyright (C) 2015-2021 Wazuh, Inc.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -9,8 +9,12 @@
  *
  * Find more information about this on the LICENSE file.
  */
-export { WazuhElasticCtrl } from './wazuh-elastic';
-export { WazuhApiCtrl } from './wazuh-api';
-export { WazuhReportingCtrl } from './wazuh-reporting';
-export { WazuhHostsCtrl } from './wazuh-hosts';
-export * from './wazuh-utils';
+
+import React from 'react';
+import ErrorBoundary from '../../error-boundary/error-boundary';
+
+export const withErrorBoundary = (WrappedComponent) => (props) => (
+  <ErrorBoundary>
+    <WrappedComponent {...props} />
+  </ErrorBoundary>
+);