+++ public/react-services/index.ts
@@ -2,3 +2,4 @@ export { GenericRequest } from './generic-request';
 export { WzRequest } from './wz-request';
 export { ErrorHandler } from './error-handler';
 export { formatUIDate } from './time-service';
+export { ErrorOrchestratorService } from './error-orchestrator/error-orchestrator.service';