+++ server/controllers/wazuh-utils/wazuh-utils.ts
@@ -11,15 +11,15 @@
  */
 
 // Require some libraries
-import { ErrorResponse } from '../lib/error-response';
-import { getConfiguration } from '../lib/get-configuration';
+import { ErrorResponse } from '../../lib/error-response';
+import { getConfiguration } from '../../lib/get-configuration';
 import { read } from 'read-last-lines';
-import { UpdateConfigurationFile } from '../lib/update-configuration';
+import { UpdateConfigurationFile } from '../../lib/update-configuration';
 import jwtDecode from 'jwt-decode';
-import { WAZUH_ROLE_ADMINISTRATOR_ID, WAZUH_DATA_LOGS_RAW_PATH } from '../../common/constants';
-import { ManageHosts } from '../lib/manage-hosts';
+import { WAZUH_ROLE_ADMINISTRATOR_ID, WAZUH_DATA_LOGS_RAW_FILENAME, WAZUH_UI_LOGS_RAW_PATH } from '../../../common/constants';
+import { ManageHosts } from '../../lib/manage-hosts';
 import { KibanaRequest, RequestHandlerContext, KibanaResponseFactory } from 'src/core/server';
-import { getCookieValueByName } from '../lib/cookie';
+import { getCookieValueByName } from '../../lib/cookie';
 
 const updateConfigurationFile = new UpdateConfigurationFile();
 
@@ -108,7 +108,7 @@ export class WazuhUtilsCtrl {
   async getAppLogs(context: RequestHandlerContext, request: KibanaRequest, response: KibanaResponseFactory) {
     try {
       const lastLogs = await read(
-        WAZUH_DATA_LOGS_RAW_PATH,
+        WAZUH_DATA_LOGS_RAW_FILENAME,
         50
       );
       const spliterLog = lastLogs.split('\n');
@@ -126,4 +126,6 @@ export class WazuhUtilsCtrl {
       return ErrorResponse(error.message || error, 3036, 500, response);
     }
   }
+
+
 }