+++ server/routes/index.ts
@@ -2,7 +2,7 @@ import { IRouter } from 'kibana/server';
 import { WazuhApiRoutes } from './wazuh-api';
 import { WazuhElasticRoutes } from "./wazuh-elastic";
 import { WazuhHostsRoutes } from "./wazuh-hosts";
-import { WazuhUtilsRoutes } from "./wazuh-utils";
+import { WazuhUtilsRoutes, UiLogsRoutes } from './wazuh-utils'
 import { WazuhReportingRoutes } from "./wazuh-reporting";
 
 export const setupRoutes = (router: IRouter) => {
@@ -11,4 +11,5 @@ export const setupRoutes = (router: IRouter) => {
     WazuhHostsRoutes(router);
     WazuhUtilsRoutes(router);
     WazuhReportingRoutes(router);
+    UiLogsRoutes(router);
 };